PYTHON-6016 Consolidate auth helper methods into auth_shared.py
