+++ pymongo/synchronous/auth.py
@@ -21,13 +21,11 @@ import hashlib
 import hmac
 import socket
 from base64 import standard_b64decode, standard_b64encode
-from collections.abc import Mapping, MutableMapping
+from collections.abc import Mapping
 from typing import (
     TYPE_CHECKING,
     Any,
     Callable,
-    Optional,
-    cast,
 )
 from urllib.parse import quote
 
@@ -35,7 +33,11 @@ from bson.binary import Binary
 from pymongo.auth_shared import (
     MongoCredential,
     _authenticate_scram_start,
+    _OIDCContext,
     _parse_scram_response,
+    _password_digest,
+    _ScramContext,
+    _X509Context,
     _xor,
 )
 from pymongo.errors import ConfigurationError, OperationFailure
@@ -43,12 +45,11 @@ from pymongo.saslprep import saslprep
 from pymongo.synchronous.auth_aws import _authenticate_aws
 from pymongo.synchronous.auth_oidc import (
     _authenticate_oidc,
-    _get_authenticator,
+    _OIDCAuthenticator,
 )
 from pymongo.synchronous.helpers import _getaddrinfo
 
 if TYPE_CHECKING:
-    from pymongo.hello import Hello
     from pymongo.synchronous.pool import Connection
 
 HAVE_KERBEROS = True
@@ -149,21 +150,6 @@ def _authenticate_scram(credentials: MongoCredential, conn: Connection, mechanis
         res = conn.command(source, cmd)
         if not res["done"]:
             raise OperationFailure("SASL conversation failed to complete.")
-
-
-def _password_digest(username: str, password: str) -> str:
-    """Get a password digest to use for authentication."""
-    if not isinstance(password, str):
-        raise TypeError("password must be an instance of str")
-    if len(password) == 0:
-        raise ValueError("password can't be empty")
-    if not isinstance(username, str):
-        raise TypeError(f"username must be an instance of str, not {type(username)}")
-
-    md5hash = hashlib.md5()  # noqa: S324
-    data = f"{username}:mongo:{password}"
-    md5hash.update(data.encode("utf-8"))
-    return md5hash.hexdigest()
 
 
 def _auth_key(nonce: str, username: str, password: str) -> str:
@@ -366,71 +352,11 @@ _AUTH_MAP: Mapping[str, Callable[..., None]] = {
 }
 
 
-class _AuthContext:
-    def __init__(self, credentials: MongoCredential, address: tuple[str, int]) -> None:
-        self.credentials = credentials
-        self.speculative_authenticate: Optional[Mapping[str, Any]] = None
-        self.address = address
-
-    @staticmethod
-    def from_credentials(
-        creds: MongoCredential, address: tuple[str, int]
-    ) -> Optional[_AuthContext]:
-        spec_cls = _SPECULATIVE_AUTH_MAP.get(creds.mechanism)
-        if spec_cls:
-            return cast(_AuthContext, spec_cls(creds, address))
-        return None
-
-    def speculate_command(self) -> Optional[MutableMapping[str, Any]]:
-        raise NotImplementedError
-
-    def parse_response(self, hello: Hello[Mapping[str, Any]]) -> None:
-        self.speculative_authenticate = hello.speculative_authenticate
-
-    def speculate_succeeded(self) -> bool:
-        return bool(self.speculative_authenticate)
-
-
-class _ScramContext(_AuthContext):
-    def __init__(
-        self, credentials: MongoCredential, address: tuple[str, int], mechanism: str
-    ) -> None:
-        super().__init__(credentials, address)
-        self.scram_data: Optional[tuple[bytes, bytes]] = None
-        self.mechanism = mechanism
-
-    def speculate_command(self) -> Optional[MutableMapping[str, Any]]:
-        nonce, first_bare, cmd = _authenticate_scram_start(self.credentials, self.mechanism)
-        # The 'db' field is included only on the speculative command.
-        cmd["db"] = self.credentials.source
-        # Save for later use.
-        self.scram_data = (nonce, first_bare)
-        return cmd
-
-
-class _X509Context(_AuthContext):
-    def speculate_command(self) -> MutableMapping[str, Any]:
-        cmd = {"authenticate": 1, "mechanism": "MONGODB-X509"}
-        if self.credentials.username is not None:
-            cmd["user"] = self.credentials.username
-        return cmd
-
-
-class _OIDCContext(_AuthContext):
-    def speculate_command(self) -> Optional[MutableMapping[str, Any]]:
-        authenticator = _get_authenticator(self.credentials, self.address)
-        cmd = authenticator.get_spec_auth_cmd()
-        if cmd is None:
-            return None
-        cmd["db"] = self.credentials.source
-        return cmd
-
-
 _SPECULATIVE_AUTH_MAP: Mapping[str, Any] = {
     "MONGODB-X509": _X509Context,
     "SCRAM-SHA-1": functools.partial(_ScramContext, mechanism="SCRAM-SHA-1"),
     "SCRAM-SHA-256": functools.partial(_ScramContext, mechanism="SCRAM-SHA-256"),
-    "MONGODB-OIDC": _OIDCContext,
+    "MONGODB-OIDC": functools.partial(_OIDCContext, authenticator_cls=_OIDCAuthenticator),
     "DEFAULT": functools.partial(_ScramContext, mechanism="SCRAM-SHA-256"),
 }
